🎨 API path 변경 및 적용
- Swagger 상세 설명 추가
- Domain path 추가
