+++ src/main/java/com/memewiki/core/domain/Heart/repository/HeartRepository.java
@@ -1,6 +1,6 @@
-package com.memewiki.core.Heart.repository;
+package com.memewiki.core.domain.Heart.repository;
 
-import com.memewiki.core.Heart.domain.Heart;
+import com.memewiki.core.domain.Heart.domain.Heart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HeartRepository extends JpaRepository<Heart, Long> {
