+++ src/main/java/com/memewiki/core/domain/meme/controller/MemeController.java
@@ -1,8 +1,9 @@
-package com.memewiki.core.meme.controller;
+package com.memewiki.core.domain.meme.controller;
 
-import com.memewiki.core.meme.response.PopularMemeResponse;
-import com.memewiki.core.meme.service.MemeService;
+import com.memewiki.core.domain.meme.response.PopularMemeResponse;
+import com.memewiki.core.domain.meme.service.MemeService;
 import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,13 +12,14 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v1/memes")
 @RequiredArgsConstructor
 @Api(tags = "meme 관련 APIs - v1")
+@RequestMapping("/api/v1/memes")
 public class MemeController {
     private final MemeService memeService;
 
     @GetMapping("/popular")
+    @ApiOperation(value = "인기있는 밈 출력 API")
     public List<PopularMemeResponse> findPopularMemes(){
         return memeService.findPopularMemes();
     }
