+++ src/main/java/com/memewiki/core/domain/meme/repository/MemeRepository.java
@@ -1,6 +1,6 @@
-package com.memewiki.core.meme.repository;
+package com.memewiki.core.domain.meme.repository;
 
-import com.memewiki.core.meme.domain.Meme;
+import com.memewiki.core.domain.meme.domain.Meme;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
