+++ src/main/java/com/memewiki/core/domain/meme/service/MemeService.java
@@ -1,7 +1,7 @@
-package com.memewiki.core.meme.service;
+package com.memewiki.core.domain.meme.service;
 
-import com.memewiki.core.meme.repository.MemeRepository;
-import com.memewiki.core.meme.response.PopularMemeResponse;
+import com.memewiki.core.domain.meme.response.PopularMemeResponse;
+import com.memewiki.core.domain.meme.repository.MemeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
